Refactored DDD concerns
